General QoL issues - Fix typing overlap between User (Type) and User (class) - Rename non-TSX files to TS - Remove unused utils
