+++ src/env.ts
@@ -1,5 +1,5 @@
-// const processEnv = typeof process !== 'undefined' ? process.env : {};
 const processEnv = process.env ?? {};
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 const injectedEnv = (window as { injectedEnv?: any }).injectedEnv || {};
 
 const env = {
