+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './dateUtils';
 export * from './validationUtils';
 export * from './stringUtils';
-export * from './apiUtils';
